vscode: General cleanup in the extension code
