+++ src/vscode/overview-view.ts
@@ -1,0 +1,96 @@
+import * as vscode from "vscode";
+
+export class OverviewViewProvider implements vscode.WebviewViewProvider {
+  public static readonly viewType = "functionGraphOverview.overview";
+  private readonly helloWorldSvg: string;
+  private _view?: vscode.WebviewView;
+
+  constructor(
+    private readonly _extensionUri: vscode.Uri,
+    helloWorldSvg: string,
+  ) {
+    this.helloWorldSvg = helloWorldSvg;
+  }
+
+  public setSVG(svg: string) {
+    if (this._view) {
+      this._view.webview.postMessage({ type: "svgImage", svg });
+    }
+  }
+
+  resolveWebviewView(
+    webviewView: vscode.WebviewView,
+    _context: vscode.WebviewViewResolveContext,
+    _token: vscode.CancellationToken,
+  ): Thenable<void> | void {
+    this._view = webviewView;
+
+    webviewView.webview.options = {
+      // Allow scripts in the webview
+      enableScripts: true,
+
+      localResourceRoots: [this._extensionUri],
+    };
+
+    webviewView.webview.html = this._getHtmlForWebview(webviewView.webview);
+  }
+  private _getHtmlForWebview(webview: vscode.Webview): string {
+    // Get the local path to main script run in the webview, then convert it to a uri we can use in the webview.
+    const scriptUri = webview.asWebviewUri(
+      vscode.Uri.joinPath(this._extensionUri, "webview-content", "main.js"),
+    );
+
+    // Do the same for the stylesheet.
+    const styleResetUri = webview.asWebviewUri(
+      vscode.Uri.joinPath(this._extensionUri, "webview-content", "reset.css"),
+    );
+    const styleVSCodeUri = webview.asWebviewUri(
+      vscode.Uri.joinPath(this._extensionUri, "webview-content", "vscode.css"),
+    );
+    const styleMainUri = webview.asWebviewUri(
+      vscode.Uri.joinPath(this._extensionUri, "webview-content", "main.css"),
+    );
+
+    // Use a nonce to only allow a specific script to be run.
+    const nonce = getNonce();
+
+    return `<!DOCTYPE html>
+              <html lang="en">
+              <head>
+                  <meta charset="UTF-8">
+  
+                  <!--
+                      Use a content security policy to only allow loading styles from our extension directory,
+                      and only allow scripts that have a specific nonce.
+                      (See the 'webview-sample' extension sample for img-src content security policy examples)
+                  -->
+                  <meta http-equiv="Content-Security-Policy" content="default-src 'none'; style-src ${webview.cspSource}; script-src 'nonce-${nonce}';">
+  
+                  <link href="${styleResetUri}" rel="stylesheet">
+                  <link href="${styleVSCodeUri}" rel="stylesheet">
+                  <link href="${styleMainUri}" rel="stylesheet">
+  
+                  <meta name="viewport" content="width=device-width, initial-scale=1.0">
+                  <title>Overview</title>
+                  <style>
+  
+                  </style>
+              </head>
+              <body>
+              <div id="overview">${this.helloWorldSvg}</div>
+  
+                  <script nonce="${nonce}" src="${scriptUri}"></script>
+              </body>
+              </html>`;
+  }
+}
+
+function getNonce() {
+  let text = "";
+  const possible =
+    "ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789";
+  for (let i = 0; i < 32; i++) {
+    text += possible.charAt(Math.floor(Math.random() * possible.length));
+  }
+  return text;
+}
