+++ webview-content/main.css
@@ -19,3 +19,7 @@ svg {
   width: 100%;
   height: 100%;
 }
+
+.highlight {
+  filter: brightness(5%);
+}
